+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unrealised]
+
+### Added
+
+- Added Anthropic Sonnet v2 model
+
 ## [4.0.7] - 2024-09-25
 
 ### Changed